feat: implement numbered/bulleted/quote list logic
