+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/mobile_toolbar_v3/aa_toolbar_item.dart
@@ -1,11 +1,12 @@
 import 'package:appflowy/generated/flowy_svgs.g.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/mobile_toolbar_v3/_align_items.dart';
 import 'package:appflowy/plugins/document/presentation/editor_plugins/mobile_toolbar_v3/_bius_items.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/mobile_toolbar_v3/_block_items.dart';
 import 'package:appflowy/plugins/document/presentation/editor_plugins/mobile_toolbar_v3/_color_item.dart';
 import 'package:appflowy/plugins/document/presentation/editor_plugins/mobile_toolbar_v3/_heading_and_text_items.dart';
 import 'package:appflowy/plugins/document/presentation/editor_plugins/mobile_toolbar_v3/util.dart';
 import 'package:appflowy/plugins/document/presentation/editor_plugins/plugins.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
-import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 
 final appflowyTextDecorationItem = AppFlowyMobileToolbarItem(
@@ -76,11 +77,13 @@ class _TextDecorationMenuState extends State<_TextDecorationMenu> {
           const ScaledVSpace(),
           Row(
             children: [
-              _BlockItems(
+              BlockItems(
                 editorState: editorState,
               ),
               const Spacer(),
-              const _AlignItems(),
+              AlignItems(
+                editorState: editorState,
+              ),
             ],
           ),
           const ScaledVSpace(),
@@ -92,67 +95,6 @@ class _TextDecorationMenuState extends State<_TextDecorationMenu> {
             ],
           ),
         ],
-      ),
-    );
-  }
-}
-
-class _BlockItems extends StatelessWidget {
-  _BlockItems({
-    required this.editorState,
-  });
-
-  final EditorState editorState;
-
-  final List<(FlowySvgData, String)> _blockItems = [
-    (FlowySvgs.m_aa_bulleted_list_s, BulletedListBlockKeys.type),
-    (FlowySvgs.m_aa_numbered_list_s, NumberedListBlockKeys.type),
-    (FlowySvgs.m_aa_quote_s, QuoteBlockKeys.type),
-    (FlowySvgs.m_aa_link_s, ToggleListBlockKeys.type),
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    return IntrinsicHeight(
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          ..._blockItems
-              .mapIndexed(
-                (index, e) => [
-                  _buildBlockItem(
-                    index,
-                    e.$1,
-                    e.$2,
-                  ),
-                  if (index != 0 || index != _blockItems.length - 1)
-                    const ScaledVerticalDivider(),
-                ],
-              )
-              .flattened,
-        ],
-      ),
-    );
-  }
-
-  Widget _buildBlockItem(
-    int index,
-    FlowySvgData icon,
-    String blockType,
-  ) {
-    return MobileToolbarItemWrapper(
-      size: const Size(62, 54),
-      enableTopLeftRadius: index == 0,
-      enableBottomLeftRadius: index == 0,
-      enableTopRightRadius: index == _blockItems.length - 1,
-      enableBottomRightRadius: index == _blockItems.length - 1,
-      showDownArrow: index == _blockItems.length - 1,
-      onTap: () {},
-      color: const Color(0xFFF2F2F7),
-      icon: icon,
-      isSelected: editorState.isBlockTypeSelected(blockType),
-      iconPadding: const EdgeInsets.symmetric(
-        vertical: 14.0,
       ),
     );
   }
@@ -218,25 +160,6 @@ class _IndentAndOutdentItems extends StatelessWidget {
           ),
         ],
       ),
-    );
-  }
-}
-
-class _AlignItems extends StatelessWidget {
-  const _AlignItems();
-
-  @override
-  Widget build(BuildContext context) {
-    return MobileToolbarItemWrapper(
-      size: const Size(82, 52),
-      onTap: () {},
-      icon: FlowySvgs.m_aa_align_left_s,
-      isSelected: false,
-      iconPadding: const EdgeInsets.symmetric(
-        vertical: 14.0,
-      ),
-      showDownArrow: true,
-      color: const Color(0xFFF2F2F7),
     );
   }
 }
